refactor(MainActivity class): button, intent, onSaveInstanceState
